#50 refactor: game-clock 参照元を新 _stores/game-clock/ へ切替
- hook (useGameClockStore) は新 _stores/game-clock/ (実装/context) を参照
- position-store.ts・intent-store.ts の GameClockStore 型は _types 集約 index 経由に統一
  (旧 _stores/game-clock-store.ts・_contexts/game-clock-store-context.tsx 自体はまだ削除しない)
