+++ src/prototypes/time-control/time-control-03/_components/action-bar/index.tsx
@@ -1,6 +1,5 @@
 import { Button } from '@/components/ui/button'
 
-import { useGameClockStore } from '../../_contexts/game-clock-store-context'
 import { usePathStore } from '../../_contexts/path-store-context'
 import { usePlannedPathStore } from '../../_contexts/planned-path-store-context'
 import { usePositionStore } from '../../_contexts/position-store-context'
@@ -9,6 +8,7 @@ import {
   DEFAULT_ACTOR_SETTINGS,
   useActorSettingsStore,
 } from '../../_stores/actor-settings'
+import { useGameClockStore } from '../../_stores/game-clock'
 import { ActorId } from '../../types'
 
 type ActionBarProps = {
