+++ src/prototypes/time-control/time-control-03/_components/action-log-panel/index.tsx
@@ -9,7 +9,7 @@ import {
 
 import { buildHistory } from '../../../time-control-02/_lib/build-history'
 import { formatCoord } from '../../../time-control-02/_lib/format-coord'
-import { useGameClockStore } from '../../_contexts/game-clock-store-context'
+import { useGameClockStore } from '../../_stores/game-clock'
 import { ActorId } from '../../types'
 
 type ActionLogPanelProps = {
